+++ proofs/is.php
@@ -55,6 +55,68 @@ return static function() {
     );
 
     yield proof(
+        'Is::string()->nonEmpty()',
+        given(
+            Set::strings()->atLeast(1),
+            Set::either(
+                Set::integers(),
+                Set::realNumbers(),
+                Set::of(
+                    true,
+                    false,
+                    null,
+                    new stdClass,
+                ),
+                Set::sequence(Set::strings()),
+            ),
+        ),
+        static function($assert, $string, $other) {
+            $constraint = Is::string()->nonEmpty();
+
+            $assert->true(
+                $constraint->asPredicate()($string),
+            );
+            $assert->same(
+                $string,
+                $constraint($string)->match(
+                    static fn($value) => $value,
+                    static fn() => null,
+                ),
+            );
+            $assert->false(
+                $constraint->asPredicate()($other),
+            );
+            $assert->same(
+                [['$', 'Value is not of type string']],
+                $constraint($other)->match(
+                    static fn() => null,
+                    static fn($failures) => $failures
+                        ->map(static fn($failure) => [
+                            $failure->path()->toString(),
+                            $failure->message(),
+                        ])
+                        ->toList(),
+                ),
+            );
+            $assert->false(
+                $constraint->asPredicate()(''),
+            );
+            $assert->same(
+                [['$', 'String cannot be empty']],
+                $constraint('')->match(
+                    static fn() => null,
+                    static fn($failures) => $failures
+                        ->map(static fn($failure) => [
+                            $failure->path()->toString(),
+                            $failure->message(),
+                        ])
+                        ->toList(),
+                ),
+            );
+        },
+    );
+
+    yield proof(
         'Is::string()->withFailure()',
         given(
             Set::strings()->atLeast(1),
@@ -126,6 +188,135 @@ return static function() {
                             $failure->message(),
                         ])
                         ->toList(),
+                ),
+            );
+        },
+    );
+
+    yield proof(
+        'Is::int()->positive()',
+        given(
+            Set::integers()->above(1),
+            Set::integers()->below(0),
+        ),
+        static function($assert, $int, $other) {
+            $constraint = Is::int()->positive();
+
+            $assert->true(
+                $constraint->asPredicate()($int),
+            );
+            $assert->same(
+                $int,
+                $constraint($int)->match(
+                    static fn($value) => $value,
+                    static fn() => null,
+                ),
+            );
+            $assert->false(
+                $constraint->asPredicate()($other),
+            );
+            $assert->same(
+                [['$', 'Integer must be above 0']],
+                $constraint($other)->match(
+                    static fn() => null,
+                    static fn($failures) => $failures
+                        ->map(static fn($failure) => [
+                            $failure->path()->toString(),
+                            $failure->message(),
+                        ])
+                        ->toList(),
+                ),
+            );
+        },
+    );
+
+    yield proof(
+        'Is::int()->negative()',
+        given(
+            Set::integers()->below(-1),
+            Set::integers()->above(0),
+        ),
+        static function($assert, $int, $other) {
+            $constraint = Is::int()->negative();
+
+            $assert->true(
+                $constraint->asPredicate()($int),
+            );
+            $assert->same(
+                $int,
+                $constraint($int)->match(
+                    static fn($value) => $value,
+                    static fn() => null,
+                ),
+            );
+            $assert->false(
+                $constraint->asPredicate()($other),
+            );
+            $assert->same(
+                [['$', 'Integer must be below 0']],
+                $constraint($other)->match(
+                    static fn() => null,
+                    static fn($failures) => $failures
+                        ->map(static fn($failure) => [
+                            $failure->path()->toString(),
+                            $failure->message(),
+                        ])
+                        ->toList(),
+                ),
+            );
+        },
+    );
+
+    yield proof(
+        'Is::int()->range()',
+        given(
+            Set::compose(
+                static fn($min, $max) => [$min, $max],
+                Set::integers(),
+                Set::integers(),
+            )
+                ->filter(static fn($bounds) => $bounds[0] < $bounds[1])
+                ->flatMap(
+                    static fn($bounds) => Set::compose(
+                        static fn($valid, $invalid) => [
+                            $bounds->unwrap()[0],
+                            $bounds->unwrap()[1],
+                            $valid,
+                            $invalid,
+                        ],
+                        Set::integers()->between(
+                            $bounds->unwrap()[0],
+                            $bounds->unwrap()[1],
+                        ),
+                        Set::either(
+                            Set::integers()->below($bounds->unwrap()[0] - 1),
+                            Set::integers()->above($bounds->unwrap()[1] + 1),
+                        ),
+                    ),
+                ),
+        ),
+        static function($assert, $in) {
+            [$min, $max, $valid, $invalid] = $in;
+
+            $constraint = Is::int()->range($min, $max);
+
+            $assert->true(
+                $constraint->asPredicate()($valid),
+            );
+            $assert->same(
+                $valid,
+                $constraint($valid)->match(
+                    static fn($value) => $value,
+                    static fn() => null,
+                ),
+            );
+            $assert->false(
+                $constraint->asPredicate()($invalid),
+            );
+            $assert->false(
+                $constraint($invalid)->match(
+                    static fn() => true,
+                    static fn() => false,
                 ),
             );
         },